backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.14,<3.15
djangorestframework-simplejwt>=5.3,<5.4
django-cors-headers>=4.3,<4.4
psycopg2-binary>=2.9,<3.0
django-environ>=0.11,<0.12
gunicorn>=21.2,<22.0
django-jazzmin>=3.0.4
whitenoise>=6.6.0
